+++ browser/data-browser/src/handlers/sideBarHandler.ts
@@ -1,53 +1,50 @@
-import { Resource, Store, urls, isString } from '@tomic/react';
+import { Resource, Store, isString, dataBrowser, core } from '@tomic/react';
 
 export function buildSideBarNewResourceHandler(store: Store) {
   // When a resource is saved add it to the parents subResources list if it's not already there.
   return async (resource: Resource) => {
-    const parentSubject = resource.get(urls.properties.parent);
+    const parentSubject = resource.get(core.properties.parent);
 
     if (!isString(parentSubject)) {
-      throw new Error(
-        `Resource doesn't have a parent: ${resource.getSubject()} `,
-      );
+      throw new Error(`Resource doesn't have a parent: ${resource.subject} `);
     }
 
-    const parent = await store.getResourceAsync(parentSubject);
-    const subResources = parent.getSubjects(urls.properties.subResources);
+    const parent = await store.getResource(parentSubject);
+    const subResources = parent.getSubjects(
+      dataBrowser.properties.subResources,
+    );
 
-    if (subResources.includes(resource.getSubject())) {
+    if (subResources.includes(resource.subject)) {
       return;
     }
 
-    await parent.pushPropVal(urls.properties.subResources, [
-      resource.getSubject(),
-    ]);
+    parent.pushPropVal(dataBrowser.properties.subResources, [resource.subject]);
 
-    await parent.save(store);
+    await parent.save();
   };
 }
 
 export function buildSideBarRemoveResourceHandler(store: Store) {
   // When a resource is deleted remove it from the parents subResources list.
   return async (resource: Resource) => {
-    const parentSubject = resource.get(urls.properties.parent);
+    const parentSubject = resource.get(core.properties.parent);
 
     if (!isString(parentSubject)) {
-      throw new Error(
-        `Resource doesn't have a parent: ${resource.getSubject()} `,
-      );
+      throw new Error(`Resource doesn't have a parent: ${resource.subject} `);
     }
 
-    const parent = await store.getResourceAsync(parentSubject);
-    const subResources = parent.getSubjects(urls.properties.subResources);
+    const parent = await store.getResource(parentSubject);
+    const subResources = parent.getSubjects(
+      dataBrowser.properties.subResources,
+    );
 
     if (subResources.length > 0) {
       await parent.set(
-        urls.properties.subResources,
-        subResources.filter(r => r !== resource.getSubject()),
-        store,
+        dataBrowser.properties.subResources,
+        subResources.filter(r => r !== resource.subject),
       );
 
-      await parent.save(store);
+      await parent.save();
     }
   };
 }