+++ browser/data-browser/src/views/OntologyPage/Class/AddPropertyButton.tsx
@@ -1,4 +1,4 @@
-import { Resource, Store, urls, useStore } from '@tomic/react';
+import { Resource, Store, core, urls, useStore } from '@tomic/react';
 import { useRef, useState } from 'react';
 import { styled } from 'styled-components';
 import { transition } from '../../../helpers/transition';
@@ -15,15 +15,20 @@ interface AddPropertyButtonProps {
 const BUTTON_WIDTH = 'calc(100% - 5.6rem + 4px)'; //Width is 100% - (2 * 1.8rem for button width) + (2rem for gaps) + (4px for borders)
 
 async function newProperty(shortname: string, parent: Resource, store: Store) {
-  const subject = `${parent.getSubject()}/property/${shortname}`;
-  const resource = store.getResourceLoading(subject, { newResource: true });
+  const subject = `${parent.subject}/property/${shortname}`;
 
-  await resource.addClasses(store, urls.classes.property);
-  await resource.set(urls.properties.shortname, shortname, store);
-  await resource.set(urls.properties.description, 'a property', store);
-  await resource.set(urls.properties.datatype, urls.datatypes.string, store);
-  await resource.set(urls.properties.parent, parent.getSubject(), store);
-  await resource.save(store);
+  const resource = await store.newResource({
+    subject,
+    parent: parent.subject,
+    isA: core.classes.property,
+    propVals: {
+      [core.properties.shortname]: shortname,
+      [core.properties.description]: 'a property',
+      [core.properties.datatype]: urls.datatypes.string,
+    },
+  });
+
+  await resource.save();
 
   return subject;
 }
@@ -51,7 +56,7 @@ export function AddPropertyButton({
         ? urls.properties.requires
         : urls.properties.recommends;
     creator.pushPropVal(creatorProp, [newValue]);
-    await creator.save(store);
+    await creator.save();
   };
 
   const handleCreateProperty = async (shortname: string) => {