+++ browser/data-browser/src/views/TablePage/PropertyForm/RelationPropertyForm.tsx
@@ -1,4 +1,4 @@
-import { urls, useStore, useString } from '@tomic/react';
+import { Datatype, core, useString } from '@tomic/react';
 import { useEffect } from 'react';
 import { styled } from 'styled-components';
 import { ResourceSelector } from '../../../components/forms/ResourceSelector';
@@ -9,15 +9,14 @@ const valueOpts = { commit: false };
 export function RelationPropertyForm({
   resource,
 }: PropertyCategoryFormProps): JSX.Element {
-  const store = useStore();
   const [classType, setClassType] = useString(
     resource,
-    urls.properties.classType,
+    core.properties.classtype,
     valueOpts,
   );
 
   useEffect(() => {
-    resource.set(urls.properties.datatype, urls.datatypes.atomicUrl, store);
+    resource.set(core.properties.datatype, Datatype.ATOMIC_URL);
   }, []);
 
   return (
@@ -25,7 +24,7 @@ export function RelationPropertyForm({
       <Label as='label'>
         <strong>Resource type:</strong>
         <ResourceSelector
-          isA={urls.classes.class}
+          isA={core.classes.class}
           value={classType}
           setSubject={setClassType}
         />